refactor: lift ensureUniqueSlug to a generic getUniqueFieldValue util
The dedupe logic isn't slug-specific — appending the first free -N suffix,
DB-checked, applies to any unique field. Moves it to utilities as
getUniqueFieldValue (field name generalized, optional locale added) so other
fields can reuse it.
