+++ packages/payload/src/fields/baseFields/slug/generateSlugBeforeDuplicate.ts
@@ -1,7 +1,7 @@
 import type { FieldHook } from '../../config/types.js'
 
+import { getUniqueFieldValue } from '../../../utilities/getUniqueFieldValue.js'
 import { markSlugForDuplicateFallback } from './duplicateContext.js'
-import { ensureUniqueSlug } from './ensureUniqueSlug.js'
 import { hasValue } from './hasValue.js'
 
 type Args = {
@@ -16,7 +16,7 @@ type Args = {
  * So the copy takes its own new id: the slug is cleared and marked so the create's `beforeChange`
  * skips source derivation and the id fallback backfills it. A required slug can't defer to that
  * fallback (pre-insert validation rejects the empty value), so it dedupes the original through
- * {@link ensureUniqueSlug} instead.
+ * {@link getUniqueFieldValue} instead.
  */
 export const generateSlugBeforeDuplicate =
   ({ name, required }: Args): FieldHook =>
@@ -27,7 +27,12 @@ export const generateSlugBeforeDuplicate =
 
     if (required) {
       return hasValue(value)
-        ? await ensureUniqueSlug({ name, collection: collection.slug, req, value: String(value) })
+        ? await getUniqueFieldValue({
+            collection: collection.slug,
+            field: name,
+            req,
+            value: String(value),
+          })
         : undefined
     }
